+++ PrayerApp/Views/Prayer/PrayerDetailPage.xaml.cs
@@ -1,3 +1,4 @@
+using PrayerApp.Helpers;
 using PrayerApp.Models;
 using PrayerApp.ViewModels;
 
@@ -27,6 +28,8 @@ public partial class PrayerDetailPage : ContentPage
 
         vm.FormResetRequested += (_, _) =>
             Dispatcher.DispatchAsync(() => TitleEntry.Focus());
+
+        EditGuardHelper.AttachEditGuardBackButton(this);
     }
 
     private void OnTagEntryCompleted(object? sender, EventArgs e)