+++ PrayerApp/Views/Tags/TagDetailPage.xaml.cs
@@ -1,3 +1,4 @@
+using PrayerApp.Helpers;
 using PrayerApp.ViewModels;
 #if ANDROID
 using Android.Views;
@@ -22,6 +23,8 @@ namespace PrayerApp.Views.Tags
         {
             InitializeComponent();
             BindingContext = vm;
+
+            EditGuardHelper.AttachEditGuardBackButton(this);
         }
 
         protected override void OnDisappearing()